+++ R/numericals7-package.R
@@ -1,4 +1,6 @@
 #' @keywords internal
+#' @importFrom Rcpp sourceCpp
+#' @useDynLib numericals7, .registration = TRUE
 "_PACKAGE"
 
 ## usethis namespace: start